adjust to latest changes in core
fix several issues around the logo and other bugs
