+++ src/main/resources/io/jenkins/plugins/customizable_header/ContextAwareLogo/config.jelly
@@ -6,9 +6,7 @@
   <f:entry field="showFolderWeather" title="${%Show the weather as logo for folders}">
     <f:checkbox/>
   </f:entry>
-  <j:if test="${it.class.name == 'hudson.model.Hudson'}">
-    <f:entry field="symbolMappingFile" title="${%Additional Mappings}">
-      <f:textbox/>
-    </f:entry>
-  </j:if>
+  <f:entry field="symbolMappingFile" title="${%Additional Mappings}">
+    <f:textbox/>
+  </f:entry>
 </j:jelly>
