+++ src/main/resources/io/jenkins/plugins/customizable_header/CustomHeaderConfiguration/config.jelly
@@ -14,6 +14,8 @@
 
           <f:dropdownDescriptorSelector field="logo" title="${%The logo to show}" descriptors="${descriptor.logoDescriptors}"/>
 
+          <f:optionalProperty field="contextAwareLogo" title="Context Aware Logo"/>
+
           <f:entry field="logoText" title="${%Text}">
               <f:textbox placeholder="Jenkins" clazz="custom-header__text-input" />
           </f:entry>
