+++ src/main/resources/io/jenkins/plugins/customizable_header/CustomHeaderDecorator/header.jelly
@@ -1,9 +1,9 @@
 <?jelly escape-by-default='true'?>
 <j:jelly xmlns:j="jelly:core">
   <link rel="stylesheet" href="${resURL}/plugin/customizable-header/css/header.css" type="text/css"/>
+  <link rel="preload" href="${rootURL}/customizable-header/theme" as="style"/>
+  <link rel="stylesheet" href="${rootURL}/customizable-header/theme" type="text/css"/>
   <j:if test="${it.enabled}">
-    <link rel="preload" href="${rootURL}/customizable-header/theme" as="style"/>
-    <link rel="stylesheet" href="${rootURL}/customizable-header/theme" type="text/css"/>
     <j:if test="${it.cssResourceUrl != ''}">
       <link rel="preload" href="${it.cssResourceUrl}" as="style"/>
       <link rel="stylesheet" href="${it.cssResourceUrl}" type="text/css"/>
