+++ src/main/resources/io/jenkins/plugins/customizable_header/UserHeader/help-links.html
@@ -1,0 +1,5 @@
+<div>
+  When links are configured, a button is generated before the logo that opens a menu with the links that open in a
+  new window. Typically, those links should point to places outside of Jenkins but which are related to this Jenkins
+  instance, e.g. Wiki pages, external documentation.
+</div>
