--- src/main/resources/io/jenkins/plugins/customizable_header/headers/ContextSelector/help-showFolderWeather.html
@@ -1,3 +1,0 @@
-<div>
-    When checked the logo will show the weather of the folder instead of the current folder icon.
-</div>
