--- src/main/resources/io/jenkins/plugins/customizable_header/headers/ContextSelector/help-showJobWeather.html
@@ -1,3 +1,0 @@
-<div>
-    When checked the logo will show the weather of the job instead of the current job status.
-</div>
