--- src/main/resources/io/jenkins/plugins/customizable_header/headers/ContextSelector/help-symbolMappingFile.html
@@ -1,9 +1,0 @@
-<div>
-    A properties file containing additional mappings for the context. They key is the class and the value is the symbol
-    that should be used. This mapping file will be consulted for symbols for context classes before trying the internal
-    mapping, which only covers a subset of contexts.<br/>
-    To find out which class you need to map create a Logger for the class
-    <em>io.jenkins.plugins.customizable_header.headers.ContextAwareHeader</em> on level <em>fine</em>. This will print
-    the list of contexts when loading a page, e.g. <em>Context: hudson.logging.LogRecorder</em>.<br/>
-    The path can be absolute or relative to JENKINS_HOME.
-</div>
