--- src/main/resources/io/jenkins/plugins/customizable_header/headers/ContextSelector/help.html
@@ -1,10 +1,0 @@
-<div>
-    Show a logo corresponding to the current context. This shows:<br/>
-    <ul>
-        <li>The job status in jobs and runs</li>
-        <li>The job weather in jobs, when show weather is checked</li>
-        <li>The folders icon in folders</li>
-        <li>On manage pages the corresponding symbol</li>
-        <li>The selected default logo when the context can't be determined</li>
-    </ul>
-</div>
