--- src/main/resources/io/jenkins/plugins/customizable_header/headers/JenkinsHeaderSelector/help.html
@@ -1,4 +1,0 @@
-<div>
-    Shows the default Jenkins header or a header from
-    another plugin.
-</div>
