--- src/main/resources/io/jenkins/plugins/customizable_header/headers/LogoHeader/headerContent.jelly
@@ -1,19 +1,0 @@
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:l="/lib/layout" xmlns:h="/lib/layout/header" xmlns:ch="/lib/customizable-header">
-  <header id="page-header" class="page-header custom-header__page">
-    <ch:links/>
-    <div class="custom-header">
-      <a class="custom-header__link" href="${rootURL}/">
-        <st:include page="logo" it="${it.logo}"/>
-        <div class="custom-header__text">${it.logoText}</div>
-      </a>
-    </div>
-    <div class="custom-header__title ">
-      <j:out value="${it.title}"/>
-    </div>
-
-    <h:searchbox/>
-    <h:login/>
-  </header>
-  <ch:systemMessages/>
-</j:jelly>
