--- src/main/resources/io/jenkins/plugins/customizable_header/headers/LogoSelector/help.html
@@ -1,3 +1,0 @@
-<div>
-    Shows the selected logo on every page.
-</div>
